+++ tests/FunctionalTest/Fixtures/Kernel.php
@@ -22,6 +22,26 @@ class Kernel extends SymfonyKernel
         yield new DIBridgeBundle();
     }
 
+    public function getProjectDir(): string
+    {
+        return __DIR__;
+    }
+
+    protected function getContainerClass(): string
+    {
+        return $this->randomName();
+    }
+
+    private function randomName(): string {
+        $characters = 'abcdefghijklmnopqrstuvwxyz';
+        $str = '';
+        for ($i = 0; $i < 10; $i++) {
+            $str .= $characters[\rand(0, \strlen($characters) - 1)];
+        }
+
+        return $str;
+    }
+
     public function registerContainerConfiguration(LoaderInterface $loader)
     {
         $loader->load(__DIR__ . '/config/' . $this->configFile);