+++ tests/FunctionalTest/Fixtures/config/di_alias.php
@@ -1,0 +1,7 @@
+<?php
+
+use function DI\get;
+
+return [
+    'class2Alias' => get('class2')
+];