+++ tests/FunctionalTest/KernelTest.php
@@ -9,9 +9,9 @@
 
 namespace Teknoo\Tests\DI\SymfonyBridge\FunctionalTest;
 
-use Teknoo\Tests\DI\SymfonyBridge\FunctionalTest\Fixtures\Class1;
 use Teknoo\Tests\DI\SymfonyBridge\FunctionalTest\Fixtures\Class2;
 use Psr\Container\ContainerInterface;
+use Teknoo\Tests\DI\SymfonyBridge\FunctionalTest\Fixtures\ContainerAwareController;
 
 /**
  * @coversNothing
@@ -25,19 +25,19 @@ class KernelTest extends AbstractFunctionalTest
         self::assertInstanceOf(ContainerInterface::class, $kernel->getContainer());
     }
 
-    public function testPhpdiShouldResolveClasses()
-    {
-        $kernel = $this->createKernel();
-
-        $object = $kernel->getContainer()->get(Class1::class);
-        self::assertInstanceOf(Class1::class, $object);
-    }
-
-    public function testSymfonyShouldResolveClasses()
+    public function testSymfonyShouldResolveClassesFromYaml()
     {
         $kernel = $this->createKernel('class2.yml');
 
         $object = $kernel->getContainer()->get('class2');
         self::assertInstanceOf(Class2::class, $object);
+    }
+
+    public function testSymfonyShouldResolveClassesFromDI()
+    {
+        $kernel = $this->createKernel('empty.yml');
+
+        $object = $kernel->getContainer()->get(ContainerAwareController::class);
+        self::assertInstanceOf(ContainerAwareController::class, $object);
     }
 }